added t/021_accents.t ; problem with text color? $pdf->content?
Hello Andrew!  Mike mwjWest@gmail.com here.  This is my first attempt
at a push, hope this is helpful; if not please forgive me. Suggestions
welcome. Drawing inspiration from Hacktober...

In the git folder, this runs with

$ perl t/021_accents.t

and saves a pdf for examination.  That pdf does not seem to
respect the color setting of text. The test also throws errors
when $pdf->content is called twice; the first time seems to empty it.