Fix color not working for text blocks
This commit firstly fixes colors not working when specified as an option
for text(). It also makes coloring more consistent. Headers and footers
now default to black, which may have previously been unpredictable based
on the previously used color. At some point an option to specify the
header/footer colors may be useful.