Lots of tweaks to make page layout cleaner and better spaced
This commit makes lots of minor tweaks to the spacing between items,
which makes the overall page layout more consistent for all content,
removing the need to make tweaks to each item's spacing. For example,
better default spacing is added above and below all header sizes.

It is a breaking change, in that PDFs previously produced with the
module will be laid out with different spacing, but as a general rule
old code can be updated by removing any manual spacing "hacks" which may
have previously been required.